# Refactor gpu interior face integration