--- a/src/riemann_solver.hpp
+++ b/src/riemann_solver.hpp
@@ -208,7 +208,6 @@ class RiemannSolver {
     }
   }
 
-
 #endif
 };