@trevilo 2022-05-03
### Performance update
I was tricked by the grvy timer output above into thinking that the restart and visualization file write step was negligible.  However, the `restart_files_hdf5` timing *only* includes the restart file write, not the visualization file write (and time required to move primitives to the host and update the pressure).  As a result, changing `flow/outputFreq` significantly affects the time spent in `solve` (at least for this case).  For instance, setting `flow/outputFreq = 100` (instead of 20 as above) and leaving everything else as before leads to the following times (still lassen, 1 mpi rank):

* On main prior to this PR:
```
-----------------------------------------------------------------------------------------------
TPS - Performance Timings:                              |      Mean      Variance       Count
--> solve               : 1.44515e+01 secs ( 75.9066 %) | [1.44515e-01  2.83485e-06        100]
--> restart_files_hdf5  : 1.40920e-02 secs (  0.0740 %) | [1.40920e-02  0.00000e+00          1]
--> GRVY_Unassigned     : 4.57292e+00 secs ( 24.0194 %)

    Total Measured Time = 1.90385e+01 secs (100.0000 %)
-----------------------------------------------------------------------------------------------
```
* On this branch:
```
-----------------------------------------------------------------------------------------------
TPS - Performance Timings:                              |      Mean      Variance       Count
--> solve               : 7.64890e+00 secs ( 62.3466 %) | [7.64890e-02  5.63179e-05        100]
--> restart_files_hdf5  : 1.54190e-02 secs (  0.1257 %) | [1.54190e-02  0.00000e+00          1]
--> GRVY_Unassigned     : 4.60403e+00 secs ( 37.5277 %)

    Total Measured Time = 1.22684e+01 secs (100.0000 %)
-----------------------------------------------------------------------------------------------
```
Now the performance improvement for `solve` is about 47%.